Extract the code to make it reusable.
