+++ es-app/src/FileData.h
@@ -67,6 +67,8 @@ public:
 	void sort(ComparisonFunction& comparator, bool ascending = true);
 	void sort(const SortType& type);
 
+	static void populateFolder(FileData* folder, const std::vector<std::string>& searchExtensions = std::vector<std::string>(), SystemData* systemData = nullptr);
+
 	MetaDataList metadata;
 
 private:
